Updated to use pandas dataframes and more
I've updated this script in three main ways:
1. Instead of opening a .csv-file in the script, it now takes a pandas
dataframe. This makes it much more flexible. I've included a script that
demonstrates how to use the function now.
2. I've added functionality to write out the value for each element in
the periodic table, instead of only being able to read it from the
colour.
3. I've rewritten the colour scheme code such that all matplotlib colour
schemes can be used. This includes divergent colour schemes designed to
center on 0, which wasn't previously possible.